refactor: factor out `VarExportSpec`
