+++ src/model/mod.rs
@@ -4,6 +4,7 @@ mod context_bag;
 mod dependency;
 mod module;
 mod rule;
+mod shared;
 mod task;
 
 pub use blockallow::BlockAllow;
@@ -12,4 +13,5 @@ pub use context_bag::{ContextBag, IsAncestor};
 pub use dependency::Dependency;
 pub use module::{CustomBuild, Module};
 pub use rule::Rule;
-pub use task::{Task, TaskError, VarExportSpec};
+pub use shared::VarExportSpec;
+pub use task::{Task, TaskError};
